+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Temporary as we are still unversioned but I figured we'd keep track of everything that changed here
 
+
+## 2016-06-30
+
+## Added
+- KdTree driven nearest neighbor
+- Personal heatmap display
+
+
 ## 2016-06-28
 
 ### Added